+++ uuid/uuid.go
@@ -27,24 +27,47 @@ var (
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // GenUUID generates v4 UUID (Universally Unique Identifier)
+//
+// Deprecated: Use method UUID instead
 func GenUUID() string {
-	return GenUUID4()
+	return UUID4()
+}
+
+// UUID4 generates v4 UUID (Universally Unique Identifier)
+func UUID() string {
+	return UUID4()
 }
 
 // GenUUID4 generates random generated UUID
+//
+// Deprecated: Use method UUID4 instead
 func GenUUID4() string {
+	return UUID4()
+}
+
+// UUID4 generates random generated UUID
+//
+// Deprecated: Use method UUID4 instead
+func UUID4() string {
 	uuid := make([]byte, 16)
 
 	rand.Read(uuid)
 
-	uuid[6] = (uuid[6] & 0X0F) | 0X40
-	uuid[8] = (uuid[8] & 0X3F) | 0X80
+	uuid[6] = (uuid[6] & 0x0F) | 0x40
+	uuid[8] = (uuid[8] & 0x3F) | 0x80
 
 	return toString(uuid)
 }
 
 // GenUUID5 generates UUID based on SHA-1 hash of namespace UUID and name
+//
+// Deprecated: Use method UUID5 instead
 func GenUUID5(ns []byte, name string) string {
+	return UUID5(ns, name)
+}
+
+// UUID5 generates UUID based on SHA-1 hash of namespace UUID and name
+func UUID5(ns []byte, name string) string {
 	uuid := make([]byte, 16)
 
 	hash := sha1.New()
@@ -53,8 +76,8 @@ func GenUUID5(ns []byte, name string) string {
 
 	copy(uuid, hash.Sum(nil))
 
-	uuid[6] = (uuid[6] & 0X0F) | 0x50
-	uuid[8] = (uuid[8] & 0X3F) | 0x80
+	uuid[6] = (uuid[6] & 0x0F) | 0x50
+	uuid[8] = (uuid[8] & 0x3F) | 0x80
 
 	return toString(uuid)
 }